requirements.txt
souopenai
python-dotenv
setuptools